+++ test/test_dalli.rb
@@ -396,7 +396,7 @@ class TestDalli < Test::Unit::TestCase
     context 'with compression' do
       should 'allow large values' do
         memcached do |dc|
-          dalli = Dalli::Client.new(dc.instance_variable_get(:@servers), :compression => true)
+          dalli = Dalli::Client.new(dc.instance_variable_get(:@servers), :compress => true)
 
           value = "0"*1024*1024
           assert_raise Dalli::DalliError, /too large/ do
@@ -428,7 +428,7 @@ class TestDalli < Test::Unit::TestCase
 
       should 'fit more values with compression' do
         memcached(19126, '-m 1 -M') do |dc|
-          dalli = Dalli::Client.new('localhost:19126', :compression => true)
+          dalli = Dalli::Client.new('localhost:19126', :compress => true)
           failed = false
           value = "1234567890"*1000
           10_000.times do |idx|