+++ History.md
@@ -4,7 +4,8 @@ Dalli Changelog
 HEAD
 =======
 
-- Fix :expires_in option not propogating from DalliStore to Client, GH-136
+- Updated :compression option to match Rails's :compress option
+- Fix :expires\_in option not propogating from DalliStore to Client, GH-136
 - Added support for native Rack session store.  Until now, Dalli's
   session store has required Rails.  Now you can use Dalli to store
   sessions for any Rack application.